Upg: remove virtuoso naming
